reorganising code
