--- get-check-result_test.go
@@ -1,58 +1,0 @@
-package checkly_test
-
-import (
-	"encoding/json"
-	"fmt"
-	"net/http"
-	"testing"
-	"time"
-
-	checkly "github.com/checkly/checkly-go-sdk"
-	"github.com/google/go-cmp/cmp"
-)
-
-func TestGetCheckResult(t *testing.T) {
-	t.Parallel()
-	resultID := "580c4e71-0109-45ba-9130-887ff01e1a7f"
-	ts := cannedResponseServer(t,
-		http.MethodGet,
-		fmt.Sprintf("/v1/check-results/%s/%s", wantCheckID, resultID),
-		validateEmptyBody,
-		http.StatusOK,
-		"GetCheckResult.json",
-	)
-	defer ts.Close()
-	client := checkly.NewClient("dummy")
-	client.HTTPClient = ts.Client()
-	client.URL = ts.URL
-	result, err := client.GetCheckResult(wantCheckID, resultID)
-	if err != nil {
-		t.Errorf("Expected no errors, got %w", err)
-	}
-	startedAt, _ := time.Parse(time.RFC3339, "2020-09-02T11:19:06.283Z")
-	stoppedAt, _ := time.Parse(time.RFC3339, "2020-09-02T11:19:06.413Z")
-	createdAt, _ := time.Parse(time.RFC3339, "2020-09-02T11:19:06.681Z")
-	expectedResult := checkly.CheckResult{
-		ID:                  "580c4e71-0109-45ba-9130-887ff01e1a7f",
-		HasErrors:           false,
-		HasFailures:         false,
-		RunLocation:         "eu-central-1",
-		StartedAt:           startedAt,
-		StoppedAt:           stoppedAt,
-		ResponseTime:        129,
-		ApiCheckResult:      &checkly.ApiCheckResult{},
-		BrowserCheckResult:  nil,
-		CheckID:             "73d29e72-6540-4bb5-967e-e07fa2c9465e",
-		CreatedAt:           createdAt,
-		Name:                "API check 1",
-		CheckRunID:          1599045546009,
-		Attempts:            1,
-		IsDegraded:          false,
-		OverMaxResponseTime: false,
-	}
-	if !cmp.Equal(expectedResult, result, nil) {
-		expected, _ := json.Marshal(expectedResult)
-		got, _ := json.Marshal(result)
-		t.Errorf("Got invalid result, expected: %s, \ngot: %s", expected, got)
-	}
-}
